style: update structure
